Refactor: 시리즈 컴포넌트 이동
