+++ src/pages/MyLolog/About/index.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import Button from '../../../components/Button';
 import styled from 'styled-components';
-import { AboutMaxWidth768px, settingMaxWidth768px } from '../../../styles/media';
+import { AboutMaxWidth768px } from '../../../styles/media';
 
 const About = () => {
   const [introduction, setIntroduction] = useState('');
